requirements.txt
pandas==1.5.3
numpy==1.23.5
boto3==1.38.11
fastparquet==2024.11.0
pyarrow==20.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.10